refactor(server-adapter): split primitive and middleware into separate files
Each adapter's module.ts is replaced with two leaner files (one concept
per file) so consumers and reviewers don't have to scroll through both
the verify primitive and its createMiddleware wrapper:

  server-adapter-web/src/
    verify-request.ts   — verifyRequest primitive (renamed from module.ts)

  server-adapter-node/src/
    verify-request.ts   — verifyRequest primitive
    middleware.ts       — createMiddleware factory

  server-adapter-socket-io/src/
    verify-socket.ts    — verifySocket primitive
    middleware.ts       — createMiddleware factory

index.ts barrels stay the public surface; no API changes.
