+++ packages/server-adapter-node/src/verify-request.ts
@@ -5,15 +5,10 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import type { IncomingMessage, ServerResponse } from 'node:http';
+import type { IncomingMessage } from 'node:http';
 import { extractBearerToken } from '@authup/server-adapter-kit';
 import type { TokenVerificationData } from '@authup/server-adapter-kit';
-import type {
-    Middleware, 
-    MiddlewareOptions, 
-    Next, 
-    VerifyRequestOptions,
-} from './types';
+import type { VerifyRequestOptions } from './types';
 
 export async function verifyRequest(
     req: IncomingMessage,
@@ -36,20 +31,4 @@ export async function verifyRequest(
     }
 
     return options.tokenVerifier.verify(token);
-}
-
-export function createMiddleware(context: MiddlewareOptions) : Middleware {
-    return async (req: IncomingMessage, _res: ServerResponse, next: Next) => {
-        try {
-            const data = await verifyRequest(req, context);
-            if (data) {
-                context.tokenVerifierHandler(req, data);
-            }
-        } catch (e) {
-            next(e as Error);
-            return;
-        }
-
-        next();
-    };
 }
