+++ packages/server-adapter-socket-io/src/index.ts
@@ -5,6 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './module';
+export * from './middleware';
 export * from './types';
+export * from './verify-socket';
