+++ packages/server-adapter-web/src/index.ts
@@ -5,5 +5,5 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './module';
 export * from './types';
+export * from './verify-request';
